version-0.03, details in readme.md

+++ server/src/main/java/org/fzu/cs03/daoyun/service/ClassMemberService.java
@@ -5,9 +5,7 @@ import com.alibaba.fastjson.JSONArray;
 import com.alibaba.fastjson.JSONObject;
 import org.fzu.cs03.daoyun.GlobalConstant;
 import org.fzu.cs03.daoyun.StatusCode;
-import org.fzu.cs03.daoyun.entity.ClassMember;
-import org.fzu.cs03.daoyun.entity.ClassMemberUpdate;
-import org.fzu.cs03.daoyun.entity.RequestUser;
+import org.fzu.cs03.daoyun.entity.*;
 import org.fzu.cs03.daoyun.exception.OrgMemberException;
 import org.fzu.cs03.daoyun.mapper.OrgMemberMapper;
 import org.fzu.cs03.daoyun.mapper.OrgnizationMapper;
@@ -47,25 +45,48 @@ public class ClassMemberService {
 
     private final Logger logger = LoggerFactory.getLogger(ClassMemberService.class);
 
-    private boolean userInOrgnization(Long userId, Long orgId){
-
+    public boolean userInOrgnization(Long userId, Long orgId){
         return orgMemberMapper.userInOrgnization(userId,orgId);
     }
-    private String createDefaultClassMemberInfo(){
+
+    private String createDefaultClassMemberInfo(Long userId){
+        AllUserInfo allUserInfo = userMapper.getAllUserInfoByUserId(userId);
         ClassMemberUpdate classMemberUpdate = new ClassMemberUpdate();
-        classMemberUpdate.setUserClassCollege("电气学院");
-        classMemberUpdate.setUserClassMajor("电力系统及其自动化");
-        classMemberUpdate.setUserClassNumber("0000");
-        classMemberUpdate.setUserClassSchool("南京航空航天大学");
+        String classSchool = allUserInfo.getSchool();
+        String classMajor = allUserInfo.getMajor();
+        String classCollege = allUserInfo.getCollege();
+        String userName = allUserInfo.getUserName();
+        String nickName = allUserInfo.getNickName();
+        String classNumber = allUserInfo.getStudentId();
+        if (nickName == null)
+            nickName = userName;
+        if (classSchool == null)
+            classSchool = "南京航空航天大学";
+        if (classMajor == null)
+            classSchool = "电力系统及其自动化";
+        if (classNumber == null)
+            classNumber = "学号未设置";
+        if (classCollege == null)
+            classCollege = "电气学院(默认)";
+
+        classMemberUpdate.setUserClassCollege(classCollege);
+        classMemberUpdate.setUserClassMajor(classMajor);
+        classMemberUpdate.setUserClassSchool(classSchool);
+        classMemberUpdate.setUserClassNickName(nickName);
+        classMemberUpdate.setUserClassNumber(classNumber);
+
         String json = JSON.toJSONString(classMemberUpdate);
         return json;
     }
 
-    public String userJoinClass(long orgCode,  HttpServletRequest request) throws Exception{
+    public String userJoinClass(Long orgCode,  HttpServletRequest request) throws Exception{
         String userName = request.getSession().getAttribute(GlobalConstant.sessionUser).toString();
         Long userId = userMapper.getUserIdByUserName(userName);
         Long orgId = orgnizationMapper.getOrgIdByOrgCode(orgCode);
-
+        if (orgId == null)
+            throw new OrgMemberException("班课不存在");
+        if (this.userInOrgnization(userId,orgId))
+            throw new OrgMemberException("用户已经存在于该班课中");
         this.addUserToClass(userId,orgId);
         return responseService.responseFactory(StatusCode.RESPONSE_OK,"加入成功");
     }
@@ -74,8 +95,12 @@ public class ClassMemberService {
         String userName = request.getSession().getAttribute(GlobalConstant.sessionUser).toString();
         Long userId = userMapper.getUserIdByUserName(userName);
         Long orgId = orgnizationMapper.getOrgIdByOrgCode(orgCode);
-        this.removeUserFromClass(userId,orgId);
+        if (orgId == null)
+            throw new OrgMemberException("班课不存在");
+        if (!this.userInOrgnization(userId,orgId))
+            throw new OrgMemberException("用户不在该班课中");
 
+        this.removeUserFromClass(userId,orgId);
         //若当前用户退出，班级为空，则删除班级
         if (orgMemberMapper.getClassMemberNumber(orgId) == 0L){
             cloudClassService.deleteCloudClass(orgId);
@@ -91,7 +116,6 @@ public class ClassMemberService {
         Long orgId = orgnizationMapper.getOrgIdByOrgCode(orgCode);
         if (orgId == null)
             throw new OrgMemberException("班课不存在");
-
         if ( ! this.userInOrgnization(userId,orgId))
             throw new OrgMemberException("用户不在该班课中");
 
@@ -104,13 +128,11 @@ public class ClassMemberService {
 
     // 获取成员信息
     public String getMembersByOrgCode(Long orgCode, HttpServletRequest request) throws Exception{
-
         String userName = request.getSession().getAttribute(GlobalConstant.sessionUser).toString();
         Long userId = userMapper.getUserIdByUserName(userName);
         Long orgId = orgnizationMapper.getOrgIdByOrgCode(orgCode);
         if (orgId == null)
             throw new OrgMemberException("班课不存在");
-
         if (!this.userInOrgnization(userId,orgId))
             throw new OrgMemberException("用户不在该班课中");
 
@@ -131,7 +153,6 @@ public class ClassMemberService {
         Long orgId = orgnizationMapper.getOrgIdByOrgCode(orgCode);
         if (orgId == null)
             throw new OrgMemberException("班课不存在");
-
         if (!this.userInOrgnization(userId,orgId))
             throw new OrgMemberException("用户不在该班课中");
 
@@ -145,34 +166,31 @@ public class ClassMemberService {
     }
 
 
+    // 以下是为service逻辑控制提供的事务级数据库操作实现，不再检查操作合理性。
+    // 因此，在service函数中需要做合法、权限检查。以及控制操作前后数据状态的合理性
     public void addUserToClass(Long userId, Long orgId) throws Exception {
-        if (orgId == null)
-            throw new OrgMemberException("班课不存在");
-        if (this.userInOrgnization(userId,orgId))
-            throw new OrgMemberException("用户已经存在于该班课中");
-
         Date date = new Date();
         SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
         String dateStr = format.format(date);
 
         //创建班级-成员的联系 , 两句最好用事务...
-        Long richId = richTextService.createRichText(this.createDefaultClassMemberInfo());
+        Long richId = richTextService.createRichText(this.createDefaultClassMemberInfo(userId));
         orgMemberMapper.addUserToOrgnization(userId,orgId,0L,richId,dateStr,false);
     }
 
+
     public void removeAllUsersFromClass(Long orgId) throws Exception{
-        if (orgId == null)
-            throw new OrgMemberException("班课不存在");
+//        if (orgId == null)
+//            throw new OrgMemberException("班课不存在");
         //删除所有班级成员
         orgMemberMapper.deleteAllUsersFromOrgnization(orgId);
     }
 
     public void removeUserFromClass(Long userId, Long orgId) throws Exception {
-        if (orgId == null)
-            throw new OrgMemberException("班课不存在");
-        if (!this.userInOrgnization(userId,orgId))
-            throw new OrgMemberException("用户不在该班课中");
-
+//        if (orgId == null)
+//            throw new OrgMemberException("班课不存在");
+//        if (!this.userInOrgnization(userId,orgId))
+//            throw new OrgMemberException("用户不在该班课中");
         //删除班级-成员的联系
         Long richTextId = orgMemberMapper.getRichTextId(userId,orgId);
         orgMemberMapper.deleteUserFromOrgnization(userId,orgId);
